Add autocomplete results to cache and fix mark bug
Instead of constructing HTML directly in the label of the awesomplete
(which caused it to add <mark></mark> on that HTML based on the input
and that could mess up the img src), now we store a JSON string in the
label and in the item function we parse that JSON string and use the
information to create the final HTML item there.

This commit also stores autocomplete queries and their results in
localStorage.